ajout de l'activité partie en cours, du boutonQuestion suivant, du Bouton résultat, du bouton Abanndoner, du boutonAccueil
