+++ bouml/quizzy-gestion/134531.bodies
@@ -1,0 +1,5 @@
+class ExampleInstrumentedTest
+!!!134531.java!!!	useAppContext() : void
+        // Context of the app under test.
+        Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
+        assertEquals("com.lasalle.quizzy", appContext.getPackageName());
