+++ bouml/quizzy-gestion/134915.bodies
@@ -1,0 +1,51 @@
+class ConnexionBluetoothClient
+!!!135811.java!!!	ConnexionBluetoothClient(inout context : Context, in adresseMAC : String)
+        this.context = context;
+        this.adresseMAC = adresseMAC;
+        bluetoothAdapter = BluetoothAdapter.getDefaultAdapter();
+!!!135939.java!!!	run() : void
+        if (bluetoothAdapter == null) {
+            Log.e(TAG, "Bluetooth non supporté sur cet appareil");
+            return;
+        }
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
+            if (ContextCompat.checkSelfPermission(context, android.Manifest.permission.BLUETOOTH_CONNECT) != PackageManager.PERMISSION_GRANTED) {
+                Log.e(TAG, "Permission BLUETOOTH_CONNECT non accordée");
+                return;
+            }
+        }
+        try {
+            BluetoothDevice device = bluetoothAdapter.getRemoteDevice(adresseMAC);
+            if (device == null) {
+                Log.e(TAG, "Appareil Bluetooth introuvable pour l'adresse : " + adresseMAC);
+                return;
+            }
+            socket = device.createRfcommSocketToServiceRecord(MY_UUID);
+            bluetoothAdapter.cancelDiscovery();
+            socket.connect();
+            outputStream = socket.getOutputStream();
+            Log.i(TAG, "Connexion réussie à : " + device.getName());
+        } catch (IOException e) {
+            Log.e(TAG, "Erreur de connexion à l'appareil : " + adresseMAC, e);
+        } catch (SecurityException e) {
+            Log.e(TAG, "Permission Bluetooth refusée", e);
+        }
+!!!136067.java!!!	envoyer(in trame : String) : void
+        try {
+            if (outputStream != null) {
+                outputStream.write(trame.getBytes());
+                outputStream.flush();
+                Log.i(TAG, "Trame envoyée : " + trame);
+            } else {
+                Log.e(TAG, "Connexion non établie. Impossible d'envoyer la trame.");
+            }
+        } catch (IOException e) {
+            Log.e(TAG, "Erreur lors de l'envoi", e);
+        }
+!!!136195.java!!!	closeConnexion() : void
+        try {
+            if (outputStream != null) outputStream.close();
+            if (socket != null) socket.close();
+        } catch (IOException e) {
+            Log.e(TAG, "Erreur lors de la fermeture", e);
+        }
