+++ bouml/quizzy-gestion/135171.bodies
@@ -1,0 +1,98 @@
+class PartieEnCours
+!!!136323.java!!!	onCreate(inout savedInstanceState : Bundle) : void
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_partie_en_cours);
+        Log.d(TAG, "onCreate()");
+
+        initialiserRessources();
+
+        Intent intent = getIntent();
+        boolean lancerTimer = intent.getBooleanExtra("lancerTimer", false);
+
+        if (lancerTimer) {
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
+                if (ContextCompat.checkSelfPermission(this, android.Manifest.permission.BLUETOOTH_CONNECT) != PackageManager.PERMISSION_GRANTED) {
+                    ActivityCompat.requestPermissions(this,
+                            new String[]{
+                                    android.Manifest.permission.BLUETOOTH_CONNECT,
+                                    android.Manifest.permission.BLUETOOTH_SCAN,
+                                    android.Manifest.permission.ACCESS_FINE_LOCATION,
+                                    android.Manifest.permission.ACCESS_COARSE_LOCATION
+                            },
+                            1
+                    );
+                }
+            }
+            String adresseMAC = "00:E0:4C:6D:20:A3";
+            connexionBluetooth = new ConnexionBluetoothClient(this, adresseMAC);
+            connexionBluetooth.start();
+
+            initialiserBluetoothEtParams(intent);
+            planifierReponseInitiale();
+        }
+!!!136451.java!!!	initialiserRessources() : void
+    this.boutonRetourAccueil = findViewById(R.id.boutonRetourAccueil);
+    this.boutonQuestionSuivante = findViewById(R.id.boutonQuestionSuivante);
+
+    boutonRetourAccueil.setOnClickListener(new View.OnClickListener() {
+        @Override
+        public void onClick(View v) {
+            Log.d(TAG, "clic boutonRetourAccueil");
+            Intent activitePrincipale = new Intent(PartieEnCours.this, Quizzy.class);
+            startActivity(activitePrincipale);
+        }
+    });
+
+        boutonQuestionSuivante.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Log.d(TAG, "clic boutonQuestionSuivante");
+                recevoirS();
+            }
+        });
+!!!136579.java!!!	initialiserBluetoothEtParams(inout intent : Intent) : void
+        themeID = intent.getIntExtra("themeID", 1);
+        tempsParQuestion = intent.getIntExtra("tempsParQuestion", 10);
+        nombreQuestions = intent.getIntExtra("nombreQuestions", 1);
+
+        connexionBluetooth = new ConnexionBluetoothClient(this, "00:E0:4C:6D:20:A3");
+        connexionBluetooth.start();
+        Log.d(TAG, "Bluetooth initialisé avec themeID=" + themeID + ", temps=" + tempsParQuestion + ", nbQuestions=" + nombreQuestions);
+!!!136707.java!!!	planifierReponseInitiale() : void
+        new Handler().postDelayed(() -> {
+            connexionBluetooth.envoyer("@@S;\n");
+            Log.d(TAG, "Révélation réponse première question");
+        }, tempsParQuestion * 1000);
+!!!136835.java!!!	recevoirS() : void
+        runOnUiThread(() -> {
+            if (questionsEnvoyees >= nombreQuestions) {
+                Log.d(TAG, "Toutes les questions ont été envoyées");
+
+                connexionBluetooth.envoyer("@@R;17;18\n");
+                try {
+                    Thread.sleep(1000);
+                } catch (InterruptedException e) {
+                    e.printStackTrace();
+                }
+                connexionBluetooth.envoyer("@@S\n");
+                return;
+            }
+
+            BaseDeDonnees baseDeDonnees = BaseDeDonnees.getInstance(getApplicationContext());
+            String trameQuestion = baseDeDonnees.getQuestionAleatoireParTheme(themeID);
+            connexionBluetooth.envoyer(trameQuestion);
+
+            Log.d(TAG, "Question envoyée : " + trameQuestion);
+            questionsEnvoyees++;
+
+            new Handler().postDelayed(() -> {
+                connexionBluetooth.envoyer("@@S\n");
+                Log.d(TAG, "Trame suivant envoyée (affichage question)");
+
+                new Handler().postDelayed(() -> {
+                    connexionBluetooth.envoyer("@@S\n");
+                    Log.d(TAG, "Trame suivant envoyée (affichage réponse)");
+                }, tempsParQuestion * 1000);
+
+            }, 1000);
+        });
