+++ bouml/quizzy-gestion/135299.bodies
@@ -1,0 +1,3 @@
+class ExampleUnitTest
+!!!136963.java!!!	addition_isCorrect() : void
+        assertEquals(4, 2 + 2);
