+++ bouml/quizzy-gestion/136106
@@ -1,7 +1,7 @@
 format 224
 "unknown" // unknown
-  revision 1
-  modified_by 42 "tv"
+  revision 2
+  modified_by 3 "btssn2"
   // class settings
   
   classdiagramsettings member_max_width 0 end
@@ -52,6 +52,34 @@ format 224
     activitydiagramsettings
      end
     class 135850 "AppCompatActivity"
+      visibility package 
+      cpp_decl ""
+      java_decl "${comment}${@}${visibility}${static}${final}${abstract}class ${name}${extends}${implements} {
+${members}}
+"
+      php_decl ""
+      python_2_2 python_decl ""
+      idl_decl ""
+      explicit_switch_type ""
+      mysql_decl ""
+      
+    end
+
+    class 134787 "SQLiteOpenHelper"
+      visibility package 
+      cpp_decl ""
+      java_decl "${comment}${@}${visibility}${static}${final}${abstract}class ${name}${extends}${implements} {
+${members}}
+"
+      php_decl ""
+      python_2_2 python_decl ""
+      idl_decl ""
+      explicit_switch_type ""
+      mysql_decl ""
+      
+    end
+
+    class 135043 "Thread"
       visibility package 
       cpp_decl ""
       java_decl "${comment}${@}${visibility}${static}${final}${abstract}class ${name}${extends}${implements} {
