+++ quizzy-gestion/app/src/main/AndroidManifest.xml
@@ -20,14 +20,18 @@
         android:theme="@style/Theme.Quizzy"
         tools:targetApi="31">
         <activity
+            android:name=".PartieEnCours"
+            android:exported="false" />
+        <activity
             android:name=".ParametresSession"
-            android:screenOrientation="landscape"
-            android:exported="true" />
+            android:exported="true"
+            android:screenOrientation="landscape" />
         <activity
             android:name=".Quizzy"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
