+++ quizzy-gestion/app/src/main/java/com/lasalle/quizzy/Quizzy.java
@@ -21,13 +21,6 @@ public class Quizzy extends AppCompatActivity
 
     private Button boutonParametrerSession;
     private Button boutonAfficherCredits;
-    private ConnexionBluetoothClient connexionBluetooth;
-    private Button boutonQuestionSuivante;
-
-    private int themeID;
-    private int tempsParQuestion;
-    private int nombreQuestions;
-    private int questionsEnvoyees = 1; // première déjà envoyée depuis ParametresSession
 
     @Override
     protected void onCreate(Bundle savedInstanceState)
@@ -35,40 +28,12 @@ public class Quizzy extends AppCompatActivity
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activite_principale);
         Log.d(TAG, "onCreate()");
-
         initialiserRessources();
-        afficherEtatPartie();
-
-        Intent intent = getIntent();
-        boolean lancerTimer = intent.getBooleanExtra("lancerTimer", false);
-
-        if (lancerTimer) {
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
-                if (ContextCompat.checkSelfPermission(this, android.Manifest.permission.BLUETOOTH_CONNECT) != PackageManager.PERMISSION_GRANTED) {
-                    ActivityCompat.requestPermissions(this,
-                            new String[] {
-                                    android.Manifest.permission.BLUETOOTH_CONNECT,
-                                    android.Manifest.permission.BLUETOOTH_SCAN,
-                                    android.Manifest.permission.ACCESS_FINE_LOCATION,
-                                    android.Manifest.permission.ACCESS_COARSE_LOCATION
-                            },
-                            1
-                    );
-                }
-            }
-            String adresseMAC = "2C:CF:67:94:F2:3D";
-            connexionBluetooth = new ConnexionBluetoothClient(this, adresseMAC);
-            connexionBluetooth.start();
-
-            initialiserBluetoothEtParams(intent);
-            planifierReponseInitiale();
-        }
     }
 
     private void initialiserRessources() {
         boutonParametrerSession = findViewById(R.id.boutonParametrerSession);
         boutonAfficherCredits = findViewById(R.id.boutonAfficherCredits);
-        boutonQuestionSuivante = findViewById(R.id.BoutonQuestionSuivante);
 
         boutonParametrerSession.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -77,81 +42,6 @@ public class Quizzy extends AppCompatActivity
                 Intent intent = new Intent(Quizzy.this, ParametresSession.class);
                 startActivity(intent);
             }
-        });
-
-        boutonQuestionSuivante.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Log.d(TAG, "clic boutonQuestionSuivante");
-                recevoirS();
-            }
-        });
-    }
-
-    private void afficherEtatPartie()
-    {
-        SharedPreferences prefs = getSharedPreferences("etat_partie", MODE_PRIVATE);
-        SharedPreferences.Editor editor = prefs.edit();
-        editor.putBoolean("partie_en_cours", false);
-        editor.apply();
-
-        TextView texteEtatPartie = findViewById(R.id.texteEtatPartie);
-        boolean partieEnCours = prefs.getBoolean("partie_en_cours", false);
-        texteEtatPartie.setText(partieEnCours ? "Partie en cours" : "Pas de partie en cours");
-    }
-
-    private void initialiserBluetoothEtParams(Intent intent)
-    {
-        themeID = intent.getIntExtra("themeID", 1);
-        tempsParQuestion = intent.getIntExtra("tempsParQuestion", 10);
-        nombreQuestions = intent.getIntExtra("nombreQuestions", 1);
-
-        connexionBluetooth = new ConnexionBluetoothClient(this, "2C:CF:67:94:F2:3D");
-        connexionBluetooth.start();
-        Log.d(TAG, "Bluetooth initialisé avec themeID=" + themeID + ", temps=" + tempsParQuestion + ", nbQuestions=" + nombreQuestions);
-    }
-
-    private void planifierReponseInitiale()
-    {
-        new Handler().postDelayed(() -> {
-            connexionBluetooth.envoyer("@@S;\n");
-            Log.d(TAG, "Révélation réponse première question");
-        }, tempsParQuestion * 1000);
-    }
-
-    public void recevoirS()
-    {
-        runOnUiThread(() -> {
-            if (questionsEnvoyees >= nombreQuestions) {
-                Log.d(TAG, "Toutes les questions ont été envoyées");
-
-                connexionBluetooth.envoyer("@@R;17;18\n");
-                try {
-                    Thread.sleep(1000);
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                }
-                connexionBluetooth.envoyer("@@S\n");
-                return;
-            }
-
-            BaseDeDonnees baseDeDonnees = BaseDeDonnees.getInstance(getApplicationContext());
-            String trameQuestion = baseDeDonnees.getQuestionAleatoireParTheme(themeID);
-            connexionBluetooth.envoyer(trameQuestion);
-
-            Log.d(TAG, "Question envoyée : " + trameQuestion);
-            questionsEnvoyees++;
-
-            new Handler().postDelayed(() -> {
-                connexionBluetooth.envoyer("@@S\n");
-                Log.d(TAG, "Trame suivant envoyée (affichage question)");
-
-                new Handler().postDelayed(() -> {
-                    connexionBluetooth.envoyer("@@S\n");
-                    Log.d(TAG, "Trame suivant envoyée (affichage réponse)");
-                }, tempsParQuestion * 1000);
-
-            }, 1000);
         });
     }
