+++ quizzy-gestion/app/src/main/res/layout/activity_partie_en_cours.xml
@@ -1,0 +1,118 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="fill_parent"
+    android:layout_height="fill_parent"
+    android:orientation="vertical"
+    tools:context=".PartieEnCours"
+    tools:ignore="MissingDefaultResource"
+    android:background="@color/bleu">
+
+    <LinearLayout
+        android:layout_width="fill_parent"
+        android:layout_height="50dp"
+        android:orientation="horizontal"
+        android:background="@color/jaune"
+        android:layout_marginBottom="100dp">
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="50dp"
+            android:text="@string/app_name"
+            android:textStyle="bold"
+            android:fontFamily="@font/open_sans"
+            android:gravity="center"
+            android:layout_weight="1" />
+
+
+    </LinearLayout>
+
+    <LinearLayout
+        android:layout_width="fill_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginEnd="400dp"
+        android:layout_marginStart="400dp"
+        android:layout_marginBottom="80dp"
+        >
+
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="50dp"
+        android:text="Une Partie est en Cours"
+        android:textStyle="bold"
+        android:fontFamily="@font/open_sans"
+        android:gravity="center"
+        android:layout_weight="1" />
+
+
+    </LinearLayout>
+
+    <LinearLayout
+        android:layout_width="fill_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginEnd="400dp"
+        android:layout_marginStart="400dp" >
+
+
+
+        <Button
+            android:id="@+id/boutonQuestionSuivante"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="Envoyer la Question Suivante"
+            android:layout_weight="1"
+            android:backgroundTint="@color/jaune"
+            />
+
+    </LinearLayout>
+
+    <LinearLayout
+        android:layout_width="fill_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginEnd="400dp"
+        android:layout_marginStart="400dp"
+        android:layout_marginBottom="20dp">
+
+        <Button
+            android:id="@+id/boutonAfficherScore"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="Afficher les Scores "
+            android:layout_weight="1"
+            android:backgroundTint="@color/jaune" />
+    </LinearLayout>
+
+    <LinearLayout
+        android:layout_width="fill_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginEnd="400dp"
+        android:layout_marginStart="400dp"
+        android:layout_marginBottom="5dp">
+
+        <Button
+            android:id="@+id/boutonAbandonnerPartie"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="Abandonner la Partie"
+            android:layout_weight="1"
+            android:backgroundTint="@color/jaune" />
+    </LinearLayout>
+
+    <LinearLayout
+        android:layout_width="fill_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginEnd="400dp"
+        android:layout_marginStart="400dp">
+
+
+        <Button
+            android:id="@+id/boutonRetourAccueil"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="Accueil"
+            android:layout_weight="1"
+            android:backgroundTint="@color/jaune" />
+    </LinearLayout>
+
+</LinearLayout>
